src/ingestion/job/requirements.txt
google-cloud-storage
google-cloud-aiplatform[generative-ai]==1.74.0
